# Optimize `CancellableContinuationImpl.invokeOnCancellation(..)` for `Segment`s