--- a/kotlinx-coroutines-core/common/src/sync/Semaphore.kt
+++ b/kotlinx-coroutines-core/common/src/sync/Semaphore.kt
@@ -185,7 +185,7 @@ internal open class SemaphoreImpl(private val permits: Int, acquiredPermits: Int
 
     private suspend fun acquireSlowPath() = suspendCancellableCoroutineReusable<Unit> sc@ { cont ->
         // Try to suspend.
-        if (addAcquireToQueue(cont as Waiter)) return@sc
+        if (addAcquireToQueue(cont)) return@sc
         // The suspension has been failed
         // due to the synchronous resumption mode.
         // Restart the whole `acquire`.