Current semaphore implementation uses `Segment`-s for storing waiting continuations. Moreover, the upcoming new channels and mutex algorithms also use segments to store waiters. When suspending, a cancellation handler should be provided via `cont.invokeOnCancellation { ... }`-- it cleans up the corresponding slot in the segment and physically removes this segment from the linked list if it becomes full of cancelled cells. However, this cancellation handler requires an allocation every time. 

To reduce the memory pressure, we can store the segment along with the slot index in `CancellableContinuationImpl` directly, as a cancellation handler instruction; thus, eliminating allocations for the corresponding cancellation handlers. For this purpose, we:
1) Allow storing `Segment` in `state` field, similarly to `CancelHandler`. On cancellation, `Segment.invokeOnCancellation(index, cause)` function is called.
2) Store the slot index in the existing `decision` integer field, extending its purpose correspondingly.

The benchmark below (see the comment) shows a significant allocation rate reduction.