@ndkoval 2023-02-15
See below the results on my laptop (MacBook Pro 16-inch, 2021, Apple M1 Max, 64 GB, OpenJDK 64-Bit Server VM Zulu19.32+13-CA).

WITHOUT the optimization:
```
Benchmark                                                                             (channelCapacity)  Mode  Cnt        Score         Error   Units
ChannelSinkBenchmark.channelPipeline                                                                N/A  avgt    5        1.375 ±       0.023   ms/op
ChannelSinkBenchmark.channelPipeline:·gc.alloc.rate.norm                                            N/A  avgt    5   668370.560 ±     256.516    B/op
ChannelSinkBenchmark.channelPipelineOneThreadLocal                                                  N/A  avgt    5        1.756 ±       0.012   ms/op
ChannelSinkBenchmark.channelPipelineOneThreadLocal:·gc.alloc.rate.norm                              N/A  avgt    5   668468.144 ±     296.998    B/op
ChannelSinkBenchmark.channelPipelineTwoThreadLocals                                                 N/A  avgt    5        2.501 ±       0.123   ms/op
ChannelSinkBenchmark.channelPipelineTwoThreadLocals:·gc.alloc.rate.norm                             N/A  avgt    5  1668726.477 ±     115.551    B/op
ChannelSinkNoAllocationsBenchmark.channelPipeline                                                   0  avgt    5        6.081 ±       0.140   ms/op
ChannelSinkNoAllocationsBenchmark.channelPipeline:·gc.alloc.rate.norm                               0  avgt    5  3426068.483 ±     354.374    B/op
ChannelSinkNoAllocationsBenchmark.channelPipeline                                                   1  avgt    5        7.167 ±       0.194   ms/op
ChannelSinkNoAllocationsBenchmark.channelPipeline:·gc.alloc.rate.norm                               1  avgt    5  3426073.447 ±     415.756    B/op
ChannelSinkNoAllocationsBenchmark.channelPipeline                                                   8  avgt    5        7.367 ±       0.148   ms/op
ChannelSinkNoAllocationsBenchmark.channelPipeline:·gc.alloc.rate.norm                               8  avgt    5  3425914.828 ±     428.938    B/op
ChannelSinkNoAllocationsBenchmark.channelPipeline                                                  32  avgt    5        7.032 ±       0.281   ms/op
ChannelSinkNoAllocationsBenchmark.channelPipeline:·gc.alloc.rate.norm                              32  avgt    5  3425325.595 ±     406.352    B/op
```

WITH the optimization:
```
Benchmark                                                                             (channelCapacity)  Mode  Cnt        Score         Error   Units
ChannelSinkBenchmark.channelPipeline                                                                N/A  avgt    5        1.248 ±       0.013   ms/op
ChannelSinkBenchmark.channelPipeline:·gc.alloc.rate.norm                                            N/A  avgt    5   488344.851 ±     129.550    B/op
ChannelSinkBenchmark.channelPipelineOneThreadLocal                                                  N/A  avgt    5        1.681 ±       0.031   ms/op
ChannelSinkBenchmark.channelPipelineOneThreadLocal:·gc.alloc.rate.norm                              N/A  avgt    5   488460.940 ±     258.884    B/op
ChannelSinkBenchmark.channelPipelineTwoThreadLocals                                                 N/A  avgt    5        2.518 ±       0.027   ms/op
ChannelSinkBenchmark.channelPipelineTwoThreadLocals:·gc.alloc.rate.norm                             N/A  avgt    5  1493167.804 ±     116.414    B/op
ChannelSinkNoAllocationsBenchmark.channelPipeline                                                   0  avgt    5        5.971 ±       0.760   ms/op
ChannelSinkNoAllocationsBenchmark.channelPipeline:·gc.alloc.rate.norm                               0  avgt    5  1025957.479 ±     274.255    B/op
ChannelSinkNoAllocationsBenchmark.channelPipeline                                                   1  avgt    5        7.455 ±       0.715   ms/op
ChannelSinkNoAllocationsBenchmark.channelPipeline:·gc.alloc.rate.norm                               1  avgt    5  1025968.018 ±     325.348    B/op
ChannelSinkNoAllocationsBenchmark.channelPipeline                                                   8  avgt    5        7.479 ±       0.666   ms/op
ChannelSinkNoAllocationsBenchmark.channelPipeline:·gc.alloc.rate.norm                               8  avgt    5  1025972.876 ±     317.774    B/op
ChannelSinkNoAllocationsBenchmark.channelPipeline                                                  32  avgt    5        7.537 ±       0.654   ms/op
ChannelSinkNoAllocationsBenchmark.channelPipeline:·gc.alloc.rate.norm                              32  avgt    5  1025967.149 ±     318.770    B/op
```