@ndkoval 2023-02-15
These benchmarks do not show performance improvement but clearly show reduced allocations.